refactor: extract main function
